# docs/openapi.json
{
  "components": {
    "schemas": {
      "AdjustmentPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdjustmentResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "AdjustmentPageResponse",
        "type": "object"
      },
      "AdjustmentRequest": {
        "additionalProperties": false,
        "properties": {
          "legs": {
            "items": {
              "$ref": "#/components/schemas/LegRequest"
            },
            "maxItems": 50,
            "title": "Legs",
            "type": "array"
          },
          "reason": {
            "maxLength": 500,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason",
          "legs"
        ],
        "title": "AdjustmentRequest",
        "type": "object"
      },
      "AdjustmentResponse": {
        "properties": {
          "approved_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "deposit_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Id"
          },
          "entry_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "manual",
              "suspense_release",
              "suspense_return"
            ],
            "title": "Kind",
            "type": "string"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/LegResponse"
            },
            "title": "Legs",
            "type": "array"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "requested_by": {
            "format": "uuid",
            "title": "Requested By",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "status",
          "kind",
          "deposit_id",
          "user_id",
          "reason",
          "legs",
          "requested_by",
          "approved_by",
          "entry_id",
          "created_at",
          "decided_at"
        ],
        "title": "AdjustmentResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SuspenseReleaseRequest": {
        "additionalProperties": false,
        "properties": {
          "deposit_id": {
            "format": "uuid",
            "title": "Deposit Id",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "title": "Reason",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "reason",
          "deposit_id",
          "user_id"
        ],
        "title": "SuspenseReleaseRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "An access token from /v1/auth/login, or an agent API key.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Corridor",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/adjustments": {
      "get": {
        "operationId": "list_adjustments_v1_admin_adjustments_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "approved",
                    "rejected"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Adjustments, newest first",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "request_adjustment_v1_admin_adjustments_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ask for a journal entry, for another admin to approve",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/adjustments/suspense-release": {
      "post": {
        "operationId": "request_suspense_release_v1_admin_adjustments_suspense_release_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspenseReleaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ask for a deposit in suspense to be credited to a user",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/adjustments/suspense-return": {
      "post": {
        "operationId": "request_suspense_return_v1_admin_adjustments_suspense_return_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspenseReturnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ask for a deposit in suspense to be booked as sent back",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/adjustments/{adjustment_id}": {
      "get": {
        "operationId": "get_adjustment_v1_admin_adjustments__adjustment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "adjustment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Adjustment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "One adjustment",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/adjustments/{adjustment_id}/approve": {
      "post": {
        "operationId": "approve_adjustment_v1_admin_adjustments__adjustment_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "adjustment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Adjustment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve another admin's adjustment, which posts it",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/adjustments/{adjustment_id}/reject": {
      "post": {
        "operationId": "reject_adjustment_v1_admin_adjustments__adjustment_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "adjustment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Adjustment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdjustmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Turn a pending adjustment down",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/deposits/suspense": {
      "get": {
        "operationId": "list_suspense_deposits_v1_admin_deposits_suspense_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuspenseDepositPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Deposits in suspense, newest first",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/outbox/dead": {
      "get": {
        "operationId": "list_dead_letters_v1_admin_outbox_dead_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadLetterPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Outbox events that ran out of attempts, newest first",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/outbox/dead/{event_id}/requeue": {
      "post": {
        "operationId": "requeue_dead_letter_v1_admin_outbox_dead__event_id__requeue_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeadLetterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Give a dead event its attempts again",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}